[reorg] Move SubstLookup and GSUB into the new layout.
